requirements.txt
aiohttp==3.5.4
aiohttp-jinja2==1.1.1
aiohttp-socks==0.2.2
aioreloader==0.2.2
async-timeout==3.0.1
attrs==19.1.0
beautifulsoup4==4.7.1
bs4==0.0.1
chardet==3.0.4
idna==2.8
Jinja2==2.10.1
MarkupSafe==1.1.1
multidict==4.5.2
PyYAML==5.1
soupsieve==1.9.1
uvloop==0.12.2
yarl==1.3.0
